+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Fixed: Inject Sentry Interceptor conditionally in backend
+* Changed: Use @sentry/node instead of nest-raven
+
 ## [v0.8.3] - 2018-11-14
 
 > Janked: Crash when Sentry is used